+++ nucleoid/lib/src/css/white_space.dart
@@ -1,0 +1,18 @@
+enum WhiteSpaceCSS { normal, nowrap, pre, preLine, preWrap }
+
+String mapperWhiteSpaceCSS(WhiteSpaceCSS value) {
+  switch (value) {
+    case WhiteSpaceCSS.normal:
+      return 'normal';
+    case WhiteSpaceCSS.nowrap:
+      return 'nowrap';
+    case WhiteSpaceCSS.pre:
+      return 'pre';
+    case WhiteSpaceCSS.preLine:
+      return 'pre-line';
+    case WhiteSpaceCSS.preWrap:
+      return 'pre-wrap';
+    default:
+      throw Exception('Invalid value WhiteSpaceCSS');
+  }
+}