+++ nucleoid/lib/src/factory/nucleoid_factory.dart
@@ -1,4 +1,8 @@
+import 'package:nucleoid/nucleoid.dart';
+
 import 'object_factory.dart';
+
+typedef FunctionTestFactory = Future<void> Function({@required bool autoCorrect});
 
 class FactoryNucleoid {
   static final FactoryNucleoid _singleton = FactoryNucleoid._internal();
@@ -11,12 +15,17 @@ class FactoryNucleoid {
 
   final Map<String, ObjectFactory> factoryMap = {};
 
-  void init(Map<String, ObjectFactory> map) {
+  void init(Map<ObjectFactory, List<String>> map) {
     map.forEach((key, value) {
-      if (factoryMap.containsKey(key)) throw Exception('FactoryNucleoid: factory "$key" init before.');
-      factoryMap[key] = value;
+      value.forEach((factory) {
+        if (factoryMap.containsKey(factory)) throw Exception('FactoryNucleoid: factory "$factory" init before.');
+        factoryMap[factory] = key;
+      });
     });
   }
+
+  Future<void> test(List<FunctionTestFactory> tests, {@required bool autoCorrect}) =>
+      Future.forEach<FunctionTestFactory>(tests, (test) => test(autoCorrect: autoCorrect));
 
   static ObjectFactory get(String key) {
     if (_singleton.factoryMap.containsKey(key)) {