+++ nucleoid_crm/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.2
+
+- update UrlModelMySQL
+
 ## 0.2.0
 
 - Check nucleoid compatible 0.2.0